NEW: find parameter index for doc

+++ chakuro-agent/demo/demo.py
@@ -7,7 +7,7 @@ from tqdm import tqdm
 import Levenshtein
 
 hello = 'world'
-np.array([1, 2, 3])
+np.array([1, 2, 3], dtype=int)
 load_iris(True)
 fft()
 print('')
